+++ Models/Nokia.cs
@@ -2,11 +2,11 @@ namespace Models;
 
 public class Nokia : Smartphone{
     // Construção da classe
-    public Nokia(string numero, string modelo, string imei, float memoriaTotal) : base(numero, modelo, imei, memoriaTotal) {
+    public Nokia(string numero, string modelo, string imei, int memoriaTotal) : base(numero, modelo, imei, memoriaTotal) {
     }
 
     // Implementação do método abstrato
-    public override void InstalarAplicativo(string nome, float tamanho) {
+    public override void InstalarAplicativo(string nome, int tamanho) {
         bool memoriaDisponivel = VerificarMemoriaDisponivel(MemoriaDisponivel);
 
         if (memoriaDisponivel) {
